# swagger/swagger.yaml
definitions:
...
  dynamic.nodeAffinity:
    properties:
      key:
        type: string
      operator:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: 请输入以 `Bearer ` 开头的 Token，例：Bearer xxxxxxxx。未列出接口请参考前端调用方法。Token在个人中心-API密钥菜单下申请。
  title: k8m API
  version: "1.0"
paths:
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/add_node_affinity/ns/{ns}/name/{name}:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: 资源组
        in: path
        name: group
        required: true
        type: string
      - description: 资源版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点亲和性配置
        in: body
        name: nodeAffinity
        required: true
        schema:
          $ref: '#/definitions/dynamic.nodeAffinity'
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 添加节点亲和性配置
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/delete_node_affinity/ns/{ns}/name/{name}:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: 资源组
        in: path
        name: group
        required: true
        type: string
      - description: 资源版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点亲和性配置
        in: body
        name: nodeAffinity
        required: true
        schema:
          $ref: '#/definitions/dynamic.nodeAffinity'
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除节点亲和性配置
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/list_node_affinity/ns/{ns}/name/{name}:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: 资源组
        in: path
        name: group
        required: true
        type: string
      - description: 资源版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: 获取节点亲和性配置
...
  /k8s/cluster/{cluster}/{kind}/group/{group}/version/{version}/update_node_affinity/ns/{ns}/name/{name}:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      - description: 资源组
        in: path
        name: group
        required: true
        type: string
      - description: 资源版本
        in: path
        name: version
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: 资源名称
        in: path
        name: name
        required: true
        type: string
      - description: 节点亲和性配置
        in: body
        name: nodeAffinity
        required: true
        schema:
          $ref: '#/definitions/dynamic.nodeAffinity'
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新节点亲和性配置
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"